Be16NDTIviolet algorithm

Applies the Be16NDTIviolet algorithm

Be16NDTIviolet(w658, w444)

Arguments

w658
numeric. Value at wavelength of 658 nm

w444
numeric. Value at wavelength of 444 nm

Value

RasterLayer or numeric

References

Beck, R.; Xu, M.; Zhan, S.; Liu, H.; Johansen, R.A.; Tong, S.; Yang, B.; Shu, S.; Wu, Q.; Wang, S.; Berling, K.; Murray, A.; Emery, E.; Reif, M.; Harwood, J.; Young, J.; Martin, M.; Stillings, G.; Stumpf, R.; Su, H.; Ye, Z.; Huang, Y. Comparison of Satellite Reflectance Algorithms for Estimating Phycocyanin Values and Cyanobacterial Total Biovolume in a Temperate Reservoir Using Coincident Hyperspectral Aircraft Imagery and Dense Coincident Surface Observations. Remote Sens. 2017, 9, 538.

See also