Run linear model (lm)

The function runs a linear model on a single independent and dependent variable and returns a data frame containing the following: r^2, p-value, slope, and intercept of the model

extract_lm(y, x, df)

Arguments

y
dependent variable

x
independent variable

df
data frame object

Value

A dataframe of the model results

References

Johansen, Richard; et al. (2018). Evaluating the portability of satellite derived chlorophyll-a algorithms for temperate inland lakes using airborne hyperspectral imagery and dense surface observations. Harmful Algae. 76. 10.1016/j.hal.2018.05.001.

R Core Team (2018). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/.

See also