Create waterquality Map with Bing basemap

This function wraps the tmap package to help users generate fast and simple data visualization of using bing basemaps along with optional geospatial objects and histogram

Map_WQ_basemap(
  WQ_extent,
  sample_points,
  map_title,
  WQ_parameter,
  points_style = "quantile",
  histogram = TRUE
)

Arguments

WQ_extent
geospatial file (vector or raster) used to extract aerial imagery from Bing basemaps

sample_points
geospatial file (.shp or .gpkg) containing sampling locations

map_title
text used to generate title of map

WQ_parameter
text referring to column heading of data being mapped (i.e. Chl-a, PC, etc.)

points_style
method to process the color scale when col is a numeric variable. Please refer to the style argument in the ?tmap::tm_raster() function for more details (Default is "quantile").

histogram
Option to add or remove a histogram of the data values. (Default is TRUE)

Value

A data visualization of the results

See also