waterquality: Satellite Derived Water Quality Detection Algorithms

The main purpose of waterquality is to quickly and easily convert satellite-based reflectance imagery into one or many well-known water quality algorithms designed for the detection of harmful algal blooms or the following pigment proxies: chlorophyll-a, blue-green algae (phycocyanin), and turbidity. Johansen et al. (2019) doi:10.21079/11681/35053 .

See also

Author

Maintainer: Richard Johansen richard.a.johansen@erdc.dren.mil (ORCID)

Authors:

Jakub Nowosad nowosad.jakub@gmail.com (ORCID)

Molly Reif Molly.K.Reif@usace.army.mil (ORCID)

Erich Emery Erich.B.Emery@usace.army.mil

Other contributors:

U.S. Army Corps of Engineers [funder]